Progress !
// 17/06/2019

:heavy_plus_sign: Add dependency
- Adding the IronPython library by [IronPython](https://ironpython.net/) (previous commit)

:heavy_minus_sign: Remove dependency
- Removing the MoonSharp library by Xanathar (previous commit)

:bug:
- Fixed a bug that caused the settings never to be updated
- Fixed a bug that caused the software to crash when your settings file was in an incorrect format, (the software was trying to load the language file with your settings which are therefore unreadable X-X, I'm stupid, I know x))
- Fixed a bug that caused the software to lock on the launch screen if the check for updates option was not enabled.
- Fixed a bug that caused the default language to always be English and not the language of your computer

:art: Improving
- Improved management of language files. Files no longer regenerate when the software is launched, unless they have been deleted. You can therefore modify them without fear of overwriting them
- Improvement of the change of command name

:zap: Performances
- Improved settings generation, which has been poorly optimized

:fire: Removed
- Deletion of the settings.txt file which has become useless

+++ Wolfy/Classes/Langs.cs
@@ -37,8 +37,17 @@ namespace Wolfy.Classes {
                     Source = new Uri("/Localization/" + _Lang, UriKind.Relative)
                 };
 
-                // TODO: Try to preserve comments & order
-                File.WriteAllText(Path.Combine(Reference.LangsPath, _Lang), XDocument.Parse(XamlWriter.Save(_Dict)).ToString());
+                // Lang path
+                string _Path = Path.Combine(Reference.LangsPath, _Lang);
+
+                // If lang file doesn't exist, create it
+                if (!File.Exists(_Path)) {
+
+                    // TODO: Try to preserve comments & order
+                    File.WriteAllText(Path.Combine(Reference.LangsPath, _Lang), XDocument.Parse(XamlWriter.Save(_Dict)).ToString());
+
+                }
+
             }
 
             #endregion
